+++ ReleaseNotes.md
@@ -6,6 +6,7 @@
 - Линукс версия
 - Добавлена возможность не компилировать нативные библиотеки, если их функции не нужны.
 - Добавлена возможность менять масштаб сцены "на лету".
+- Добавлена возможность менять скорость воспроизведения в оффлайн режиме. 
 # Ломающие изменения:
 - Обновлённый протокол protobuf требует поменять названия переменных в коде и перекомпилировать.
 
@@ -17,5 +18,6 @@
 - Linux version
 - Now you don't need to build all native libraries if you think you don't need them.
 - Now you can change scene scale on the fly.
+- Now you can change play speed in offline mode.
 # Breaking changes
 - You need to update names in your code and recompile to use updated protobuf protocol.